requirements.txt
mysql-connector-python==8.0.32
playwright==1.31.1
lxml